+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Orbs/RedOrb.cs
@@ -15,7 +15,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Orbs
     public class RedOrb : Orb
     {
         /// <summary>The object ID of the red orb.</summary>
-        public override int ObjectID => (int)OrbType.RedOrb;
+        public override short ObjectID => (short)(int)OrbType.RedOrb;
 
         /// <summary>Initializes a new instance of the <seealso cref="RedOrb"/> class.</summary>
         public RedOrb() : base() { }
