+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Orbs/YellowOrb.cs
@@ -15,7 +15,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Orbs
     public class YellowOrb : Orb
     {
         /// <summary>The object ID of the yellow orb.</summary>
-        public override int ObjectID => (int)OrbType.YellowOrb;
+        public override short ObjectID => (short)(int)OrbType.YellowOrb;
 
         /// <summary>Initializes a new instance of the <seealso cref="YellowOrb"/> class.</summary>
         public YellowOrb() : base() { }
