+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Portals/SpeedPortals/FasterSpeedPortal.cs
@@ -14,7 +14,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Port
     public class FasterSpeedPortal : SpeedPortal, IHasCheckedProperty
     {
         /// <summary>The object ID of the faster speed portal.</summary>
-        public override int ObjectID => (int)PortalType.FasterSpeed;
+        public override short ObjectID => (short)(int)PortalType.FasterSpeed;
 
         /// <summary>The speed this speed portal sets.</summary>
         public override Speed Speed => Speed.Faster;
