+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Portals/YellowGravityPortal.cs
@@ -11,7 +11,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Port
     public class YellowGravityPortal : Portal
     {
         /// <summary>The object ID of the yellow gravity portal.</summary>
-        public override int ObjectID => (int)PortalType.YellowGravity;
+        public override short ObjectID => (short)(int)PortalType.YellowGravity;
 
         /// <summary>Initializes a new instance of the <seealso cref="YellowGravityPortal"/> class.</summary>
         public YellowGravityPortal() : base() { }
